requirements.txt
numpy==1.24.4
openai==1.59.6
pyperclip==1.9.0
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
sounddevice==0.5.1
